Loads of changes to improve behaviour with python 2.7 and in 64-bit mode.
A rather significant change is that Foundation.NSRect and Quartz.CGRect
are now aliases (because of the structure of PyObjC both are Foundation.NSRect).
Likewise for NSSize and NSPoint.

I've also added objc.array_property, objc.set_property and objc.dict_property,
but haven't fleshed out unittests for these yet.