This fixes a number of small issues:
* The wrappers for the NSBitmapImagerep initializers were incomplete, now both methods
  that take bitmap plane arguments are implemented and work.

* Fix issue with cfmessage port wrappers

* Remove debug print statuements

All of this is tested on python 2.7 and will be backported to 2.3.1 once I've finished
testing.

With this patch all tests pass, although the cfrunloop and cfusernotification tests still
fail most of the time when running the entire suite. This is probably related to how I set up
the loop and clean up resources, the tests run fine when running those test modules seperately.