Use local testrunner instead of the one from setuptools
Setup.py now contains a fully fleshed out test runner, instead
of relying on the default 'test' command in setuptools/distribute.

This has three advantages:
1) Better control over how tests are run
2) Improved isolation from the installed version of pyobjc-core
3) The test commands now prints a summary of the test results
   that is easily parsed by other scripts.

KeyValueCoding was changed to ensure 2to3 for python 3.2 doesn't
break the file.  Test_archive_python was changed to disable a
number of stdlib tests that are invalid for pyobjc.