Fix compiler warnings for clang
This patch fixes a number of small issues in the PyObjC source code
and ensures that we compile basicly warnings-free with clang
(Xcode 4.2, -Wall).

Als add "-(BOOL)supportsWeakPointers { return YES; }" to the OC_Python*
classes, this makes sure we play nice with ARC (our release implementation
calls the super one after grabbing the python GIL and should therefore
be safe in ARC mode).