Add PyObjC_FAST_BUT_INEXACT define to pyobjc.h to switch between two options for method resolution
The option is there to make it easier to performance test post variants,
the option is off by default for now because this closer matches the 2.x
semantics.

First tests seem to indicate that enabling the option results in a faster
bridge, but I've been testing on a machine that was not idle at the time
and the difference is not very great.