Improve docstring for objc.native_selector and objc.function.
Both now show the prototype for the object, which gives nicer
introspection and should make pydoc output somewhat usefull.

The docstring is generated using a python function, the docstring
should not be on the fast path and by using a python function
it is easy to adjust the contents of the docstring.

TODO:
- python function is not yet finished, and there is more
  information available than just the prototype.
- for some reason pydoc for an objc.function shows information
  on that class instead of the instance.