Clean up the pyobjc-api.h header
This mostly removes a large fraction of the definitions
in the API structure, and I hope to further reduce the
API in the near future.

An open issue w.r.t. to the API is the test module
in Modules/objc/test/ctest.m: that module is used to
provide C-level unittests for parts of the API and
needs functions I don't want to expose. I'd also
like to add more tests, which likely results in exposing
more of PyObjC's internals. This may result in moving
that extension into the _objc extension and exposing it
as a (private) attribute for the unittests.