Simplify some code in class-builder.m.
In particular the list of methods to be added to the first
generation Python class is now an actual list, instead of being
'unrolled' in code. This makes it easier to verify that
the two different mechanisms behave the same, and also makes
it easier to change the list.

The cleanup also fixes some small inconsistencies.