Tweak the build procedure to avoid building pyobc-core multipel times
See issue #21

This patch adds symlinks to pyobjc-api.h and pyobjc-compat.h to all
framework wrappers that have an Exension module. The setup.py file
removes the setup_requires on pyobjc-core when those headers are
present (the setup_requires is still present).

This does not yet close the issue because I haven't tested if this
completely fixes the problem.